+++ src/lib/url-sync.js
@@ -127,7 +127,7 @@ class URLSync {
     // like hierarchicalFacet.rootPath are then triggering a default refinement that would
     // be not present if it was not going trough the SearchParameters constructor
     this.originalConfig = algoliasearchHelper(
-      { addAlgoliaAgent() {} },
+      {},
       currentConfiguration.index,
       currentConfiguration
     ).state;